+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Upcoming
 
-### 🔄 Changed
+### 🐞 Fixed
+- Various events did not trigger channel update callback
 
 # [2.6.5](https://github.com/GetStream/stream-chat-swift/releases/tag/2.6.5)
 _April 12, 2021_